requirements.txt
torch
transformers
datasets
trl
peft
bitsandbytes
accelerate
pyyaml
optimum
# auto-gptq
tensorboard
